Only reference picture base dir in bootstrap.rs
